requirements.txt
requests
colorama
phonenumbers
python-whois
dnspython
beautifulsoup4
fake-useragent
